7 Deploy Redis in standalone mode
Redis is deployed in standalone mode with one primary node and two
replicas.Deployment in standalone mode was necessary as the library used
in the application was not able to connect to a cluster.With the
standalone mode the application is always connecting to the primary
node, i.e. the read/write node.