Decrease deployment limits
The service only has two instances provisioned with one task running in
each.Withe a maximum of 200% and minimum of 100%, on redeployment of a
service we are stuck in a state where not enough resources are available
to complete the deployment.

With this changes, 1 task/instance is drained from incumbent deployment
while the another task is started in the new deployment, and the same is
done for the second task after the first has been done.