#12 Change sp auth tasks network mode to brige
The AWSVPC network mode gives the instance container running in the task
an elatic network interface with their own ip address.This has the
advantage of allowing one to deploy multiple instance containers in an
ec2 instance without networking clashes.However, instance containers
deployed in an awsvpc mode WILL NOT have access to the internet without
a NAT Gateway.

It is necessary to change the target type to instance because, as stated
above, awsvpc gives instance containers their own private ip address,
which the load balancer uses to detect them when distributing traffic.
With a target type of instance the load balancer users the instance id.