Update task defintion env variables
Update the environment variables to there are no references to specific
technologies, i.e. Postgres and Redis, and the missing verification url
and reset password url are added.

Finally, the database url is removed, to bring it closer to how db
connection creds are provided in other services and local env.